Email Connector Phase 2 development
